requirements.txt
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
autopep8==2.0.4
binaryornot==0.4.4
boto3==1.34.19
botocore==1.34.19
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
cryptography==41.0.7
dj-database-url==2.1.0
Django==5.0
django-admin-honeypot-updated-2021==1.2.0
django-appconf==1.0.6
django-browser-reload==1.12.1
django-compressor==4.4
django-csp==3.7
django-embed-video==1.4.9
django-environ==0.11.2
django-ipware==6.0.3
django-js-asset==2.2.0
django-social-share==2.3.0
django-storages==1.14.2
django-tailwind==3.8.0
gunicorn==21.2.0
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-ipware==2.0.1
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
rich==13.7.0
rjsmin==1.2.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.6
whitenoise==6.6.0
